refactor: Change FormInput component location
